requirements.txt
Django==5.2.5
djangorestframework==3.16.1
gunicorn==23.0.0
python-dotenv==1.1.1
